Commit: search use case interactor test
